Move bin.pdb into bin.formats.pdb
