+++ libr/bin/Makefile
@@ -29,7 +29,6 @@ OBJS+=mangling/objc.o mangling/cxx.o mangling/msvc.o
 OBJS+=mangling/swift.o mangling/swift-sd.o
 OBJS+=mangling/java.o
 OBJS+=mangling/rust.o
-OBJS+=pdb/pdb_downloader.o pdb/omap.o pdb/stream_pe.o pdb/gdata.o
-OBJS+=pdb/fpo.o pdb/dbi.o pdb/tpi.o pdb/stream_file.o pdb/pdb.o
 
+include $(TOP)/libr/bin/format/pdb/deps.mk
 include $(TOP)/libr/rules.mk
