+++ libr/bin/format/pdb/deps.mk
@@ -1,0 +1,3 @@
+
+OBJS+=format/pdb/pdb_downloader.o format/pdb/omap.o format/pdb/stream_pe.o format/pdb/gdata.o
+OBJS+=format/pdb/fpo.o format/pdb/dbi.o format/pdb/tpi.o format/pdb/stream_file.o format/pdb/pdb.o
