+++ libr/bin/format/pdb/gdata.c
