+++ libr/bin/format/pdb/gdata.h
