+++ libr/bin/format/pdb/stream_file.h
