+++ libr/bin/format/pdb/stream_pe.h
