+++ libr/bin/meson.build
@@ -121,6 +121,16 @@ r_bin_sources = [
   join_paths('format','wasm','wasm.c'),
   join_paths('format','zimg','zimg.c'),
 
+  join_paths('format/pdb','dbi.c'),
+  join_paths('format/pdb','fpo.c'),
+  join_paths('format/pdb','gdata.c'),
+  join_paths('format/pdb','omap.c'),
+  join_paths('format/pdb','pdb.c'),
+  join_paths('format/pdb','pdb_downloader.c'),
+  join_paths('format/pdb','stream_file.c'),
+  join_paths('format/pdb','stream_pe.c'),
+  join_paths('format/pdb','tpi.c')
+
   join_paths('mangling','cxx.c'),
   join_paths('mangling','cxx','cp-demangle.c'),
   join_paths('mangling','demangler.c'),
@@ -131,16 +141,6 @@ r_bin_sources = [
   join_paths('mangling','objc.c'),
   join_paths('mangling','swift-sd.c'),
   join_paths('mangling','swift.c'),
-
-  join_paths('pdb','dbi.c'),
-  join_paths('pdb','fpo.c'),
-  join_paths('pdb','gdata.c'),
-  join_paths('pdb','omap.c'),
-  join_paths('pdb','pdb.c'),
-  join_paths('pdb','pdb_downloader.c'),
-  join_paths('pdb','stream_file.c'),
-  join_paths('pdb','stream_pe.c'),
-  join_paths('pdb','tpi.c')
 ]
 
 r_bin_sources += r_bin_d_sources
